+++ tests/backends/local/events_test.py
@@ -17,7 +17,9 @@ class LocalBackendEventsTest(unittest.TestCase):
     config = {
         'local': {
             'media_dir': path_to_data_dir(''),
+            'data_dir': path_to_data_dir(''),
             'playlists_dir': b'',
+            'library': 'json',
         }
     }