+++ tests/backends/local/playlists_test.py
@@ -20,6 +20,8 @@ class LocalPlaylistsProviderTest(unittest.TestCase):
     config = {
         'local': {
             'media_dir': path_to_data_dir(''),
+            'data_dir': path_to_data_dir(''),
+            'library': 'json',
         }
     }